requirements.txt
asgiref==3.9.1
astroid==3.3.11
colorama==0.4.6
diff-match-patch==20241021
dill==0.4.0
Django==5.2.5
django-ckeditor==6.7.3
django-import-export==4.3.9
django-js-asset==3.1.2
et_xmlfile==2.0.0
isort==6.0.1
mccabe==0.7.0
mysqlclient==2.2.7
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pylint==3.3.8
pylint-celery==0.3
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tomlkit==0.13.3
tzdata==2025.2
